requirements.txt
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
